best effort matching and more configuration options
We implemented a best effort matching strategy, which detects every language with its configured language detections and calculates a score for each match. The language with the highest score is returned. The best effort strategy is required for situations like javascript and typescript, both are using the same interpreter, but they use differnt file extensions.

We also implemented a ContentTypeDetector, which can be configured with a builder. The builder allows the configuration of a custom set language type detectors and the used the matching strategy.

The ContentTypes utils used now the best effort matching strategy for detections with path and content. The strategy for the detection of path only, still uses first match strategy.